+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Display number of generated Monero addresses in instance report.
 - Allow posts with attachment and no text.
 - Add `updated` property to edited `Note` objects.
+- Implement `/api/v1/statuses/{status_id}/source` API endpoint.
 
 ### Changed